+++ example_application/plugins/capacitor-rate-app.js
@@ -1,0 +1,7 @@
+import { registerPlugin } from './@capacitor/core.js';
+
+const RateApp = registerPlugin('RateApp', {
+    web: () => import('./common/web-d47d0e1c.js').then(m => new m.RateAppWeb()),
+});
+
+export { RateApp };