+++ example_application/plugins/common/web-d47d0e1c.js
@@ -1,0 +1,9 @@
+import { WebPlugin } from '../@capacitor/core.js';
+
+class RateAppWeb extends WebPlugin {
+    async requestReview() {
+        return;
+    }
+}
+
+export { RateAppWeb };